Interrupts are now working.
GDT:
*Fixed the argument sizes in gdt.c declaration of the Assembly symbol. The
second argument (which is the size) is now a unsigned 16 bit integer.

IDT:
*Rewritten based on the GDT module (that i know worked).
*Tested with IRQ 0 (division by 0 exception) and it works fine.

Kernel:
*Added a 'naked' interrupt function for testing that prints a message
on the screen when it is called
*It now loads the IDT and adds an ISR on the 0 exception (division by 0).
After that it tests it by performing 15/0. The ISR gets called successfuly.

Makefile:
*Changed the compiler from GCC to Clang as it supports the 'naked' attribute for
functions (needed for ISR functions).

+++ kernel/gdt/gdt.c
@@ -21,16 +21,16 @@ static _gdt_entry_struct_t _global_descriptor_table[3];
 
 /*
  * The external symbol of the label that contains the code that will load the GDT
- * Takes the address and size of the GDT as arguments
+ * Takes the address and size of the GDT as arguments (in that order)
  */
-extern void gdt_asm_load_global_descriptor_table(uint32_t, uint32_t);
+extern void gdt_asm_load_global_descriptor_table(uint32_t, uint16_t);
 
 /*
  * Load the GDT into the CPU using the 'LGDT' instruction
  */
 void gdt_load_global_descriptor_table(void)
 {
-    gdt_asm_load_global_descriptor_table(_global_descriptor_table, sizeof(_global_descriptor_table));
+    gdt_asm_load_global_descriptor_table( _global_descriptor_table, sizeof(_global_descriptor_table));
 }
 
 /*
